requirements.txt
Django==1.4
PIL==1.1.7
argparse==1.2.1
bleach==1.1.2
distribute==0.6.24
django-debug-toolbar==0.9.4
django-rosetta==0.6.8
html5lib==0.95
sorl-thumbnail==11.12
wsgiref==0.1.2
